ADD : Battery UI to fix

+++ Assets/Scripts/UI/HUD/Player UI/BatteryScript.cs
@@ -3,11 +3,11 @@ using UnityEngine.UI;
 
 public class BatteryScript : MonoBehaviour
 {
-    public Slider slider;
+    public Slider slider; 
     public Gradient gradient;
     public Image fill;
 
-    public void SetMaxBattery(int battery)
+    public void SetMaxBattery(float battery)
     {
         slider.maxValue = battery;
         slider.value = battery;
@@ -15,7 +15,7 @@ public class BatteryScript : MonoBehaviour
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetBattery(int battery)
+    public void SetBattery(float battery)
     {
         slider.value = battery;
